server/pyproject.toml
[project]
name = "stutterease-server"
version = "0.1.0"
description = "API that handles all the business logic for StutterEase."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "faster-whisper>=1.1.1",
    "ffmpeg>=1.4",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=8.4.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
